+++ src/main/java/com/rackspace/salus/zone_watcher/config/ServicesProperties.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.rackspace.salus.zone_mgmt.config;
+package com.rackspace.salus.zone_watcher.config;
 
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
