+++ src/main/java/com/rackspace/salus/zone_watcher/handler/ExpiringZoneEventProcessor.java
@@ -14,38 +14,36 @@
  * limitations under the License.
  */
 
-package com.rackspace.salus.zone_mgmt.handler;
+package com.rackspace.salus.zone_watcher.handler;
 
-import static com.rackspace.salus.telemetry.etcd.types.Keys.PTN_ZONE_ACTIVE;
-import static com.rackspace.salus.telemetry.etcd.types.Keys.TRACKING_KEY_ZONE_ACTIVE;
+import static com.rackspace.salus.telemetry.etcd.types.Keys.PTN_ZONE_EXPIRING;
+import static com.rackspace.salus.telemetry.etcd.types.Keys.TRACKING_KEY_ZONE_EXPIRING;
 
-import com.coreos.jetcd.Watch.Watcher;
 import com.coreos.jetcd.common.exception.ClosedClientException;
 import com.coreos.jetcd.common.exception.ClosedWatcherException;
 import com.coreos.jetcd.data.ByteSequence;
 import com.coreos.jetcd.watch.WatchEvent;
 import com.coreos.jetcd.watch.WatchResponse;
 import com.rackspace.salus.telemetry.etcd.services.ZoneStorage;
-import com.rackspace.salus.zone_mgmt.services.ZoneStorageListener;
+import com.rackspace.salus.zone_watcher.services.ZoneStorageListener;
 import com.rackspace.salus.telemetry.etcd.types.ResolvedZone;
-import java.util.Optional;
 import java.util.regex.Matcher;
 import lombok.extern.slf4j.Slf4j;
 
 /**
  * This is used to handle any etcd watcher events that get triggered relating
- * to the active zone key prefix.
+ * to the expiring zone key prefix.
  */
 @Slf4j
-public class ActiveZoneEventProcessor extends ZoneEventProcessor {
+public class ExpiringZoneEventProcessor extends ZoneEventProcessor {
 
-  public ActiveZoneEventProcessor(ZoneStorageListener listener, ZoneStorage zoneStorage) {
+  public ExpiringZoneEventProcessor(ZoneStorageListener listener, ZoneStorage zoneStorage) {
     super(listener, zoneStorage);
   }
 
   @Override
   public void run() {
-    final ByteSequence trackingKey = ByteSequence.fromString(TRACKING_KEY_ZONE_ACTIVE);
+    final ByteSequence trackingKey = ByteSequence.fromString(TRACKING_KEY_ZONE_EXPIRING);
 
     while (zoneStorage.isRunning()) {
       try {
@@ -55,10 +53,10 @@ public class ActiveZoneEventProcessor extends ZoneEventProcessor {
           for (WatchEvent event : response.getEvents()) {
 
             final String keyStr = event.getKeyValue().getKey().toStringUtf8();
-            final Matcher matcher = PTN_ZONE_ACTIVE.matcher(keyStr);
+            final Matcher matcher = PTN_ZONE_EXPIRING.matcher(keyStr);
 
             if (!matcher.matches()) {
-              log.warn("Unable to parse active event key={}", keyStr);
+              log.warn("Unable to parse expiring event key={}", keyStr);
               continue;
             }
 
@@ -70,23 +68,18 @@ public class ActiveZoneEventProcessor extends ZoneEventProcessor {
 
             switch (event.getEventType()) {
               case PUT:
-                // Only act on this when a new key is created
-                // not when the bound monitor count is updated for the key.
-                if (event.getPrevKV().getVersion() == 0) {
-                  listener.handleActiveEnvoyConnection(resolvedZone, resourceId);
-                }
+                // no action needed
                 break;
               case DELETE:
-                Optional<String> envoyId = zoneStorage.getEnvoyIdForResource(resolvedZone, resourceId).join();
-                if (!envoyId.isPresent()) {
-                  log.warn("No expected envoyId found for disconnected resource {}", resourceId);
-                  break;
+                // if the lease isn't expired, it means the poller came back up and
+                // the expiring entry was removed by the listener before the entry could time out.
+                if (zoneStorage.isLeaseExpired(event)) {
+                  listener.handleExpiredEnvoy(
+                      resolvedZone, resourceId, event.getPrevKV().getValue().toStringUtf8());
                 }
-                // create an expiring entry with the poller timeout.
-                listener.handleActiveEnvoyDisconnection(resolvedZone, resourceId, envoyId.get());
                 break;
               case UNRECOGNIZED:
-                log.warn("Unknown active watcher event seen by zone watcher");
+                log.warn("Unknown expiring watcher event seen by zone watcher");
                 break;
             }
           }
@@ -101,7 +94,6 @@ public class ActiveZoneEventProcessor extends ZoneEventProcessor {
         return;
       }
     }
-
     listener.handleExpectedZoneWatcherClosed(null);
   }
 }
