+++ src/main/java/com/rackspace/salus/zone_watcher/handler/ZoneEventProcessor.java
@@ -14,11 +14,11 @@
  * limitations under the License.
  */
 
-package com.rackspace.salus.zone_mgmt.handler;
+package com.rackspace.salus.zone_watcher.handler;
 
 import com.coreos.jetcd.Watch.Watcher;
 import com.rackspace.salus.telemetry.etcd.services.ZoneStorage;
-import com.rackspace.salus.zone_mgmt.services.ZoneStorageListener;
+import com.rackspace.salus.zone_watcher.services.ZoneStorageListener;
 
 public abstract class ZoneEventProcessor implements Runnable {
   Watcher zoneWatcher;
