+++ src/main/java/com/rackspace/salus/zone_watcher/services/ZoneStorageListener.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.rackspace.salus.zone_mgmt.services;
+package com.rackspace.salus.zone_watcher.services;
 
 import com.coreos.jetcd.common.exception.EtcdException;
 import com.rackspace.salus.telemetry.etcd.types.ResolvedZone;
