+++ src/main/resources/application-dev.yml
@@ -6,5 +6,5 @@ server:
   port: 8084
 logging:
   level:
-    com.rackspace.salus.zone_mgmt: debug
+    com.rackspace.salus.zone_watcher: debug
     com.rackspace.salus.telemetry.etcd: debug
